requirements.txt
absl-py==0.12.0
astor==0.8.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
clang==5.0
cloudpickle==2.2.1
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
dill==0.3.4
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.3
grpcio==1.48.2
h5py==3.1.0
idna==3.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
networkx==2.5.1
numpy>=1.17.3,<2.0
oauthlib==3.2.2
OpenEXR==1.3.9
opt-einsum==3.3.0
Pillow==8.4.0
promise==2.3
protobuf==3.19.6
psutil==7.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9.1
scipy>=1.5.4,<1.25.0
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-addons==0.14.0
tensorflow-datasets==4.5.2
tensorflow-estimator==2.6.0
tensorflow-graphics==2021.12.3
tensorflow-metadata==1.2.0
tensorflow-probability==0.7.0
termcolor==1.1.0
tf-slim==1.1.0
tqdm==4.64.1
trimesh==3.23.5
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.20
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
